git-svn-id: http://10.10.170.94/program@8 de0f5a65-c6c9-414c-ab6f-108cdf64cfb5

+++ matlab/data.analysis/bot.m
@@ -6,11 +6,10 @@ m=zeros(length(roi),imNumber);
 for i=1:length(roi)
     mIm=reshape(im(roi(i).ygrid,roi(i).xgrid,:),length(roi(i).ygrid),length(roi(i).xgrid)*imNumber);
     mBW=repmat(roi(i).BW(roi(i).ygrid,roi(i).xgrid),1,imNumber);
-    mIm=mIm.*uint16(mBW);
+    mIm=mIm.*double(mBW);
     xIm=reshape(mIm,length(roi(i).ygrid).*length(roi(i).xgrid),imNumber);
     m(i,:)=sum(xIm,1)./sum(roi(i).BW(:));
 end
-
-m=m';
+ m=m';
 return
